refactor(ui): split disclosure anchor coverage
